+++ tests/php/bootstrap.php
@@ -6,22 +6,19 @@
  */
 
 $_tests_dir = getenv( 'WP_TESTS_DIR' );
-if ( ! $_tests_dir ) {
-	$_tests_dir = '/tmp/wordpress-tests-lib';
-}
 
 define( 'PHPUNIT_SRM_TESTSUITE', true );
 
 // Give access to tests_add_filter() function.
 require_once $_tests_dir . '/includes/functions.php';
 
-/**
- * Manually load the plugin being tested.
- */
-function _manually_load_plugin() {
-	require dirname( dirname( dirname( __FILE__ ) ) ) . '/safe-redirect-manager.php';
-}
-tests_add_filter( 'muplugins_loaded', '_manually_load_plugin' );
+// Activate the plugin.
+tests_add_filter(
+	'muplugins_loaded',
+	static function (): void {
+		require dirname( __DIR__, 2 ) . '/safe-redirect-manager.php';
+	}
+);
 
 // Start up the WP testing environment.
 require $_tests_dir . '/includes/bootstrap.php';